+++ src/zuh_r.h
@@ -32,5 +32,6 @@ SEXP C_zuh_node_attr(SEXP ptr, SEXP ids, SEXP name, SEXP dflt);
 SEXP C_zuh_node_attrs(SEXP ptr, SEXP ids);
 SEXP C_zuh_node_text(SEXP ptr, SEXP ids, SEXP recursive);
 SEXP C_zuh_node_serialize(SEXP ptr, SEXP ids, SEXP outer);
+SEXP C_zuh_select(SEXP ptr, SEXP ids, SEXP css, SEXP mode, SEXP work_limit);
 
 #endif